# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для аутентификации пользователей",
        "title": "Medods Test Task API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth": {
            "get": {
                "description": "Генерирует access и refresh токены для пользователя по guid. Токены возвращаются в httpOnly cookie.",
                "tags": [
                    "auth"
                ],
                "summary": "Получить access и refresh токены",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1e44baa9-e04b-4739-89f3-3d86b9a272ce",
                        "description": "GUID пользователя",
                        "name": "guid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Токены успешно созданы",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "access_token=eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJzZXNzaW9uX2lkIjoiOTJiYmRkMTUtNmE3Ny00MWM2LTg2MzAtN2NkYWVjYmRmOGVjIiwidXNlcl9hZ2VudCI6IlBvc3RtYW5SdW50aW1lLzcuNDQuMSIsImlwIjoiMTcyLjE4LjAuMSIsImV4cCI6MTc1MTg2NDc1NCwiaWF0IjoxNzUxODYzODU0fQ.IdFs5M0cYIa1w7iBqvj5sgtvB8ufuHGpsGVUACnJxCT4MVg2nmFPPxBOamZU4nLVQOjFy9bmBqKzcNxdAb5z5Q;refresh_token=FbI-aH-FhxcYoZaZ6mdb9pBDzi--OcKL7sJT5Du04Xk"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "description": "Обновляет access и refresh токены с помощью текущего refresh токена из cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление токенов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токены в cookie (необходимы refresh_token и access_token)",
                        "name": "Cookie",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Токены успешно обновлены",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "refresh_token=\u003cновый_refresh_token\u003e"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные или отсутствующие токены",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    },
                    "401": {
                        "description": "Невалидные или просроченные токены",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/revoke": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Отзывает текущую сессию пользователя по JWT токену. Удаляет сессию из системы.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Отзыв токенов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен доступа в формате: Bearer \u003cJWT\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пустой ответ при успешном отзыве",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ: отсутствует/неверный заголовок Authorization, невалидный токен",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при отзыве токена",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает UUID пользователя по JWT токену из заголовка Authorization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Получить UUID пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен доступа в формате: Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UUID пользователя в формате строки",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}